+++ test/acceptance/db.go
@@ -1,0 +1,10 @@
+package acceptance
+
+import (
+	"database/sql"
+)
+
+// DbContext can be used to expose the database from the application.
+type DbContext struct {
+	DB *sql.DB
+}