+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - gRPC Dial function
-
+- DB acceptance context
 
 ## [0.15.1] - 2018-01-24